Move V1->V2 stats conversion to DataSourceV2Relation.transformV1Stats
